hir_def::attrs

Struct AttrFlags

pub struct AttrFlags(<AttrFlags as PublicFlags>::Internal);

Tuple Fields
0: <AttrFlags as PublicFlags>::Internal
Implementations

impl AttrFlags

pub const RUST_ANALYZER_SKIP: Self

pub const LANG_ITEM: Self

pub const HAS_DOC_ALIASES: Self

pub const HAS_DOC_KEYWORD: Self

pub const IS_DOC_NOTABLE_TRAIT: Self

pub const IS_DOC_HIDDEN: Self

pub const RUSTC_HAS_INCOHERENT_INHERENT_IMPLS: Self

pub const RUSTC_ALLOW_INCOHERENT_IMPL: Self

pub const FUNDAMENTAL: Self

pub const RUSTC_SKIP_ARRAY_DURING_METHOD_DISPATCH: Self

pub const RUSTC_SKIP_BOXED_SLICE_DURING_METHOD_DISPATCH: Self

pub const HAS_REPR: Self

pub const HAS_TARGET_FEATURE: Self

pub const RUSTC_DEPRECATED_SAFE_2024: Self

pub const HAS_LEGACY_CONST_GENERICS: Self

pub const NO_MANGLE: Self

pub const NON_EXHAUSTIVE: Self

pub const RUSTC_RESERVATION_IMPL: Self

pub const RUSTC_CONST_PANIC_STR: Self

pub const MAY_DANGLE: Self

pub const RUSTC_INTRINSIC: Self

pub const RUSTC_SAFE_INTRINSIC: Self

pub const RUSTC_INTRINSIC_MUST_BE_OVERRIDDEN: Self

pub const RUSTC_ALLOCATOR: Self

pub const RUSTC_DEALLOCATOR: Self

pub const RUSTC_REALLOCATOR: Self

pub const RUSTC_ALLOCATOR_ZEROED: Self

pub const IS_UNSTABLE: Self

pub const IS_IGNORE: Self

pub const IS_BENCH: Self

pub const IS_TEST: Self

pub const IS_EXPORT_NAME_MAIN: Self

pub const IS_MACRO_EXPORT: Self

pub const IS_NO_STD: Self

pub const IS_DERIVE_OR_BUILTIN_MACRO: Self

pub const IS_DEPRECATED: Self

pub const HAS_PATH: Self

pub const HAS_CFG: Self

pub const COMPLETE_IGNORE_FLYIMPORT: Self

pub const COMPLETE_IGNORE_FLYIMPORT_METHODS: Self

pub const COMPLETE_IGNORE_METHODS: Self

pub const RUSTC_LAYOUT_SCALAR_VALID_RANGE: Self

pub const RUSTC_PAREN_SUGAR: Self

pub const RUSTC_COINDUCTIVE: Self

pub const RUSTC_FORCE_INLINE: Self

pub const IS_POINTEE: Self

pub const MACRO_STYLE_BRACES: Self

pub const MACRO_STYLE_BRACKETS: Self

pub const MACRO_STYLE_PARENTHESES: Self

impl AttrFlags

pub const fn empty() -> Self

Get a flags value with all bits unset.

pub const fn all() -> Self

Get a flags value with all known bits set.

pub const fn bits(&self) -> u64

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u64) -> Option<Self>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u64) -> Self

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u64) -> Self

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<Self>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: Self) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: Self) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: Self)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: Self)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: Self)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: Self, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: Self) -> Self

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: Self) -> Self

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: Self) -> Self

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: Self) -> Self

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> Self

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl AttrFlags

pub const fn iter(&self) -> Iter<AttrFlags>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<AttrFlags>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.

impl AttrFlags

pub fn query(db: &dyn DefDatabase, owner: AttrDefId) -> AttrFlags

pub fn query_field(db: &dyn DefDatabase, field: FieldId) -> AttrFlags

pub fn query_generic_params( db: &dyn DefDatabase, def: GenericDefId, ) -> &(ArenaMap<LocalLifetimeParamId, AttrFlags>, ArenaMap<LocalTypeOrConstParamId, AttrFlags>)

pub fn query_lifetime_param( db: &dyn DefDatabase, owner: LifetimeParamId, ) -> AttrFlags

pub fn query_type_or_const_param( db: &dyn DefDatabase, owner: TypeOrConstParamId, ) -> AttrFlags

pub(crate) fn is_cfg_enabled_for( owner: &dyn HasAttrs, cfg_options: &CfgOptions, ) -> Result<(), CfgExpr>

pub fn lang_item(db: &dyn DefDatabase, owner: AttrDefId) -> Option<Symbol>

pub fn lang_item_with_attrs( self, db: &dyn DefDatabase, owner: AttrDefId, ) -> Option<Symbol>

pub fn repr(db: &dyn DefDatabase, owner: AdtId) -> Option<ReprOptions>

pub(crate) fn legacy_const_generic_indices( db: &dyn DefDatabase, owner: FunctionId, ) -> &Option<Box<[u32]>>

Call this only if there are legacy const generics, to save memory.

pub fn doc_html_root_url(db: &dyn DefDatabase, krate: Crate) -> &Option<SmolStr>

pub fn target_features( db: &dyn DefDatabase, owner: FunctionId, ) -> &FxHashSet<Symbol>

pub fn rustc_layout_scalar_valid_range( db: &dyn DefDatabase, owner: AdtId, ) -> RustcLayoutScalarValidRange

pub fn doc_aliases( self, db: &dyn DefDatabase, owner: Either<AttrDefId, FieldId>, ) -> &[Symbol]

pub fn cfgs( self, db: &dyn DefDatabase, owner: Either<AttrDefId, FieldId>, ) -> Option<&CfgExpr>

pub fn doc_keyword(db: &dyn DefDatabase, owner: ModuleId) -> Option<Symbol>

pub fn docs(db: &dyn DefDatabase, owner: AttrDefId) -> &Option<Box<Docs>>

pub fn field_docs(db: &dyn DefDatabase, field: FieldId) -> Option<&Docs>

pub fn derive_info(db: &dyn DefDatabase, owner: MacroId) -> Option<&DeriveInfo>