Struct hir_def::body::Body

pub struct Body {
    pub exprs: Arena<Expr>,
    pub pats: Arena<Pat>,
    pub or_pats: FxHashMap<PatId, Arc<[PatId]>>,
    pub labels: Arena<Label>,
    pub params: Vec<PatId>,
    pub body_expr: ExprId,
    /* private fields */
}

The body of an item (function, const etc.).

Fields
exprs: Arena<Expr>pats: Arena<Pat>or_pats: FxHashMap<PatId, Arc<[PatId]>>labels: Arena<Label>params: Vec<PatId>

The patterns for the function’s parameters. While the parameter types are part of the function signature, the patterns are not (they don’t change the external type of the function).

If this Body is for the body of a constant, this will just be empty.

body_expr: ExprId

The ExprId of the actual body expression.

Implementations

impl Body

pub fn blocks<'a>(
&'a self,
db: &'a dyn DefDatabase
) -> impl Iterator<Item = (BlockId, Arc<DefMap>)> + '_

Returns an iterator over all block expressions in this body that define inner items.

pub fn pattern_representative(&self, pat: PatId) -> PatId

pub fn ident_patterns_for<'slf>(&'slf self, pat: &'slf PatId) -> &'slf [PatId]

Retrieves all ident patterns this pattern shares the ident with.

pub fn pretty_print(&self, db: &dyn DefDatabase, owner: DefWithBodyId) -> String