Struct hir_def::db::InternMacroRulesQuery

pub struct InternMacroRulesQuery;

Implementations

impl InternMacroRulesQuery

pub fn in_db(self, db: &dyn InternDatabase) -> QueryTable<'_, Self>

Get access to extra methods pertaining to this query. You can also use it to invoke this query, though it’s more common to use the trait method on the database itself.

impl InternMacroRulesQuery

pub fn in_db_mut(self, db: &mut dyn InternDatabase) -> QueryTableMut<'_, Self>

Like in_db, but gives access to methods for setting the value of an input. Not applicable to derived queries.

Threads, cancellation, and blocking

Mutating the value of a query cannot be done while there are still other queries executing. If you are using your database within a single thread, this is not a problem: you only have &self access to the database, but this method requires &mut self.

However, if you have used snapshot to create other threads, then attempts to set will block the current thread until those snapshots are dropped (usually when those threads complete). This also implies that if you create a snapshot but do not send it to another thread, then invoking set will deadlock.

Before blocking, the thread that is attempting to set will also set a cancellation flag. This will cause any query invocations in other threads to unwind with a Cancelled sentinel value and eventually let the set succeed once all threads have unwound past the salsa invocation.

If your query implementations are performing expensive operations without invoking another query, you can also use the Runtime::unwind_if_cancelled method to check for an ongoing cancellation and bring those operations to a close, thus allowing the set to succeed. Otherwise, long-running computations may lead to “starvation”, meaning that the thread attempting to set has to wait a long, long time. =)