hir_def::dyn_map::keys
Struct AstPtrPolicy

pub struct AstPtrPolicy<AST, ID> { /* private fields */ }

XXX: AST Nodes and SyntaxNodes have identity equality semantics: nodes are equal if they point to exactly the same object.

In general, we do not guarantee that we have exactly one instance of a syntax tree for each file. We probably should add such guarantee, but, for the time being, we will use identity-less AstPtr comparison.