hir_def

Enum ModuleDefId

pub enum ModuleDefId {
    ModuleId(ModuleId),
    FunctionId(FunctionId),
    AdtId(AdtId),
    EnumVariantId(EnumVariantId),
    ConstId(ConstId),
    StaticId(StaticId),
    TraitId(TraitId),
    TypeAliasId(TypeAliasId),
    BuiltinType(BuiltinType),
    MacroId(MacroId),
}

The defs which can be visible in the module.

Variants

ModuleId(ModuleId)

FunctionId(FunctionId)

AdtId(AdtId)

EnumVariantId(EnumVariantId)

ConstId(ConstId)

StaticId(StaticId)

TraitId(TraitId)

TypeAliasId(TypeAliasId)

BuiltinType(BuiltinType)

MacroId(MacroId)

Implementations

impl ModuleDefId

pub fn module(&self, db: &dyn DefDatabase) -> Option<ModuleId>

Returns the module containing self (or self, if self is itself a module).

Returns None if self refers to a primitive type.