hir_def::expr_store::body

Struct BodySourceMap

pub struct BodySourceMap {
    pub self_param: Option<InFile<SelfParamPtr>>,
    pub store: ExpressionStoreSourceMap,
}

An item body together with the mapping from syntax nodes to HIR expression IDs. This is needed to go from e.g. a position in a file to the HIR expression containing it; but for type inference etc., we want to operate on a structure that is agnostic to the actual positions of expressions in the file, so that we don’t recompute types whenever some whitespace is typed.

One complication here is that, due to macro expansion, a single Body might be spread across several files. So, for each ExprId and PatId, we record both the HirFileId and the position inside the file. However, we only store AST -> ExprId mapping for non-macro files, as it is not clear how to handle this properly for macros.

Fields
self_param: Option<InFile<SelfParamPtr>>store: ExpressionStoreSourceMap
Implementations

impl BodySourceMap

pub fn self_param_syntax(&self) -> Option<InFile<SelfParamPtr>>