hir_def
Module find_path

An algorithm to find a path to refer to a certain item.

Enums

PrefixKind

Functions

find_path

Find a path that can be used to refer to a certain item. This can depend on from where you’re referring to the item, hence the from parameter.