hir_def::find_path
Function find_path

pub fn find_path(
    db: &dyn DefDatabase,
    item: ItemInNs,
    from: ModuleId,
    prefix_kind: PrefixKind,
    ignore_local_imports: bool,
    cfg: ImportPathConfig,
) -> Option<ModPath>

Find a path that can be used to refer to a certain item. This can depend on from where you’re referring to the item, hence the from parameter.