Module hir_def::generics

Many kinds of items or constructs can have generic parameters: functions, structs, impls, traits, etc. This module provides a common HIR for these generic parameters. See also the Generics type and the generics_of query in rustc.

Structs

ConstParamData

Data about a generic const parameter (to a function, struct, impl, …).

GenericParams

Data about the generic parameters of a function, struct, impl, etc.

LifetimeParamData

Data about a generic lifetime parameter (to a function, struct, impl, …).

TypeParamData

Data about a generic type parameter (to a function, struct, impl, …).

Enums

GenericParamData

GenericParamDataRef

TypeOrConstParamData

TypeParamProvenance

WherePredicate

A single predicate from a where clause, i.e. where Type: Trait. Combined where clauses like where T: Foo + Bar are turned into multiple of these. It might still result in multiple actual predicates though, because of associated type bindings like Iterator<Item = u32>.

WherePredicateTypeTarget