hir_def::hir

Module generics

Pre-type IR item generics

Structs

ConstParamData

Data about a generic const parameter (to a function, struct, impl, …).

GenericParams

Data about the generic parameters of a function, struct, impl, etc.

LifetimeParamData

Data about a generic lifetime parameter (to a function, struct, impl, …).

TypeParamData

Data about a generic type parameter (to a function, struct, impl, …).

Enums

GenericParamData

GenericParamDataRef

TypeOrConstParamData

TypeParamProvenance

WherePredicate

A single predicate from a where clause, i.e. where Type: Trait. Combined where clauses like where T: Foo + Bar are turned into multiple of these. It might still result in multiple actual predicates though, because of associated type bindings like Iterator<Item = u32>.

Type Aliases

LocalLifetimeParamId

LocalTypeOrConstParamId