hir_def::hir
Module type_ref

HIR for references to types. Paths in these are not yet resolved. They can be directly created from an ast::TypeRef, without further queries.

Structs

ArrayType

FnType

LifetimeRef

PathId

A TypeRefId that is guaranteed to always be TypeRef::Path. We use this for things like impl’s trait, that are always paths but need to be traced back to source code.

RefType

TraitRef

TypesMap

TypesSourceMap

Enums

ConstRef

LiteralConstRef

A literal constant value

Mutability

Rawness

TraitBoundModifier

A modifier on a bound, currently this is only used for ?Sized, where the modifier is Maybe.

TypeBound

TypeRef

Compare ty::Ty

UseArgRef

Type Aliases

TypePtr

TypeRefId

TypeSource