hir_def

Module lang_item

Collects lang items: items marked with #[lang = "..."] attribute.

This attribute to tell the compiler about semi built-in std library features, such as Fn family of traits.

Structs

LangItems

Enums

GenericRequirement

LangItem

A representation of all the valid language items in Rust.

LangItemTarget

Functions

crate_lang_items

Salsa query. This will look for lang items in a specific crate.

lang_item

Salsa query. Look for a lang item, starting from the specified crate and recursively traversing its dependencies.