Enum hir_def::lang_item::LangItem

pub enum LangItem {

    Sized,
    Unsize,
    StructuralPeq,
    StructuralTeq,
    Copy,
    Clone,
    Sync,
    DiscriminantKind,
    Discriminant,
    PointeeTrait,
    Metadata,
    DynMetadata,
    Freeze,
    Drop,
    Destruct,
    CoerceUnsized,
    DispatchFromDyn,
    TransmuteOpts,
    TransmuteTrait,
    Add,
    Sub,
    Mul,
    Div,
    Rem,
    Neg,
    Not,
    BitXor,
    BitAnd,
    BitOr,
    Shl,
    Shr,
    AddAssign,
    SubAssign,
    MulAssign,
    DivAssign,
    RemAssign,
    BitXorAssign,
    BitAndAssign,
    BitOrAssign,
    ShlAssign,
    ShrAssign,
    Index,
    IndexMut,
    UnsafeCell,
    VaList,
    Deref,
    DerefMut,
    DerefTarget,
    Receiver,
    Fn,
    FnMut,
    FnOnce,
    FnOnceOutput,
    Future,
    GeneratorState,
    Generator,
    Unpin,
    Pin,
    PartialEq,
    PartialOrd,
    Panic,
    PanicNounwind,
    PanicFmt,
    PanicDisplay,
    ConstPanicFmt,
    PanicBoundsCheck,
    PanicInfo,
    PanicLocation,
    PanicImpl,
    PanicCannotUnwind,
    BeginPanic,
    ExchangeMalloc,
    BoxFree,
    DropInPlace,
    AllocLayout,
    Start,
    EhPersonality,
    EhCatchTypeinfo,
    OwnedBox,
    PhantomData,
    ManuallyDrop,
    MaybeUninit,
    AlignOffset,
    Termination,
    Try,
    Tuple,
    SliceLen,
    TryTraitFromResidual,
    TryTraitFromOutput,
    TryTraitBranch,
    TryTraitFromYeet,
    PointerSized,
    Poll,
    PollReady,
    PollPending,
    ResumeTy,
    IdentityFuture,
    GetContext,
    Context,
    FuturePoll,
    FromFrom,
    OptionSome,
    OptionNone,
    ResultOk,
    ResultErr,
    ControlFlowContinue,
    ControlFlowBreak,
    IntoFutureIntoFuture,
    IntoIterIntoIter,
    IteratorNext,
    PinNewUnchecked,
    RangeFrom,
    RangeFull,
    RangeInclusiveStruct,
    RangeInclusiveNew,
    Range,
    RangeToInclusive,
    RangeTo,
    String,

}

A representation of all the valid language items in Rust.

Variants

Sized

The sized lang item.

Unsize

The unsize lang item.

StructuralPeq

The structural_peq lang item. Trait injected by #[derive(PartialEq)], (i.e. “Partial EQ”).

StructuralTeq

The structural_teq lang item. Trait injected by #[derive(Eq)], (i.e. “Total EQ”; no, I will not apologize).

Copy

The copy lang item.

Clone

The clone lang item.

Sync

The sync lang item.

DiscriminantKind

The discriminant_kind lang item.

Discriminant

The discriminant_type lang item. The associated item of the [DiscriminantKind] trait.

PointeeTrait

The pointee_trait lang item.

Metadata

The metadata_type lang item.

DynMetadata

The dyn_metadata lang item.

Freeze

The freeze lang item.

Drop

The drop lang item.

Destruct

The destruct lang item.

CoerceUnsized

The coerce_unsized lang item.

DispatchFromDyn

The dispatch_from_dyn lang item.

TransmuteOpts

The transmute_opts lang item.

TransmuteTrait

The transmute_trait lang item.

Add

The add lang item.

Sub

The sub lang item.

Mul

The mul lang item.

Div

The div lang item.

Rem

The rem lang item.

Neg

The neg lang item.

Not

The not lang item.

BitXor

The bitxor lang item.

BitAnd

The bitand lang item.

BitOr

The bitor lang item.

Shl

The shl lang item.

Shr

The shr lang item.

AddAssign

The add_assign lang item.

SubAssign

The sub_assign lang item.

MulAssign

The mul_assign lang item.

DivAssign

The div_assign lang item.

RemAssign

The rem_assign lang item.

BitXorAssign

The bitxor_assign lang item.

BitAndAssign

The bitand_assign lang item.

BitOrAssign

The bitor_assign lang item.

ShlAssign

The shl_assign lang item.

ShrAssign

The shr_assign lang item.

Index

The index lang item.

IndexMut

The index_mut lang item.

UnsafeCell

The unsafe_cell lang item.

VaList

The va_list lang item.

Deref

The deref lang item.

DerefMut

The deref_mut lang item.

DerefTarget

The deref_target lang item.

Receiver

The receiver lang item.

Fn

The fn lang item.

FnMut

The fn_mut lang item.

FnOnce

The fn_once lang item.

FnOnceOutput

The fn_once_output lang item.

Future

The future_trait lang item.

GeneratorState

The generator_state lang item.

Generator

The generator lang item.

Unpin

The unpin lang item.

Pin

The pin lang item.

PartialEq

The eq lang item.

PartialOrd

The partial_ord lang item.

Panic

The panic lang item.

PanicNounwind

The panic_nounwind lang item.

PanicFmt

The panic_fmt lang item.

PanicDisplay

The panic_display lang item.

ConstPanicFmt

The const_panic_fmt lang item.

PanicBoundsCheck

The panic_bounds_check lang item.

PanicInfo

The panic_info lang item.

PanicLocation

The panic_location lang item.

PanicImpl

The panic_impl lang item.

PanicCannotUnwind

The panic_cannot_unwind lang item.

BeginPanic

The begin_panic lang item. libstd panic entry point. Necessary for const eval to be able to catch it

ExchangeMalloc

The exchange_malloc lang item.

BoxFree

The box_free lang item.

DropInPlace

The drop_in_place lang item.

AllocLayout

The alloc_layout lang item.

Start

The start lang item.

EhPersonality

The eh_personality lang item.

EhCatchTypeinfo

The eh_catch_typeinfo lang item.

OwnedBox

The owned_box lang item.

PhantomData

The phantom_data lang item.

ManuallyDrop

The manually_drop lang item.

MaybeUninit

The maybe_uninit lang item.

AlignOffset

The align_offset lang item. Align offset for stride != 1; must not panic.

Termination

The termination lang item.

Try

The Try lang item.

Tuple

The tuple_trait lang item.

SliceLen

The slice_len_fn lang item.

TryTraitFromResidual

The from_residual lang item.

TryTraitFromOutput

The from_output lang item.

TryTraitBranch

The branch lang item.

TryTraitFromYeet

The from_yeet lang item.

PointerSized

The pointer_sized lang item.

Poll

The Poll lang item.

PollReady

The Ready lang item.

PollPending

The Pending lang item.

ResumeTy

The ResumeTy lang item.

IdentityFuture

The identity_future lang item.

GetContext

The get_context lang item.

Context

The Context lang item.

FuturePoll

The poll lang item.

FromFrom

The from lang item.

OptionSome

The Some lang item.

OptionNone

The None lang item.

ResultOk

The Ok lang item.

ResultErr

The Err lang item.

ControlFlowContinue

The Continue lang item.

ControlFlowBreak

The Break lang item.

IntoFutureIntoFuture

The into_future lang item.

IntoIterIntoIter

The into_iter lang item.

IteratorNext

The next lang item.

PinNewUnchecked

The new_unchecked lang item.

RangeFrom

The RangeFrom lang item.

RangeFull

The RangeFull lang item.

RangeInclusiveStruct

The RangeInclusive lang item.

RangeInclusiveNew

The range_inclusive_new lang item.

Range

The Range lang item.

RangeToInclusive

The RangeToInclusive lang item.

RangeTo

The RangeTo lang item.

String

The String lang item.

Implementations

impl LangItem

pub fn from_name(name: &Name) -> Option<Self>

Opposite of LangItem::name

pub fn path(&self, db: &dyn DefDatabase, start_crate: CrateId) -> Option<Path>

impl LangItem

pub fn name(self) -> SmolStr

pub fn from_str(name: &str) -> Option<Self>

Opposite of LangItem::name