hir_def::lang_item

Struct LangItems

pub struct LangItems {

    pub Sized: Option<TraitId>,
    pub MetaSized: Option<TraitId>,
    pub PointeeSized: Option<TraitId>,
    pub Unsize: Option<TraitId>,
    pub StructuralPeq: Option<TraitId>,
    pub StructuralTeq: Option<TraitId>,
    pub Copy: Option<TraitId>,
    pub Clone: Option<TraitId>,
    pub TrivialClone: Option<TraitId>,
    pub Sync: Option<TraitId>,
    pub DiscriminantKind: Option<TraitId>,
    pub Discriminant: Option<TypeAliasId>,
    pub PointeeTrait: Option<TraitId>,
    pub Metadata: Option<TypeAliasId>,
    pub DynMetadata: Option<StructId>,
    pub Freeze: Option<TraitId>,
    pub FnPtrTrait: Option<TraitId>,
    pub FnPtrAddr: Option<FunctionId>,
    pub Drop: Option<TraitId>,
    pub Destruct: Option<TraitId>,
    pub CoerceUnsized: Option<TraitId>,
    pub DispatchFromDyn: Option<TraitId>,
    pub TransmuteOpts: Option<StructId>,
    pub TransmuteTrait: Option<TraitId>,
    pub Add: Option<TraitId>,
    pub Sub: Option<TraitId>,
    pub Mul: Option<TraitId>,
    pub Div: Option<TraitId>,
    pub Rem: Option<TraitId>,
    pub Neg: Option<TraitId>,
    pub Not: Option<TraitId>,
    pub BitXor: Option<TraitId>,
    pub BitAnd: Option<TraitId>,
    pub BitOr: Option<TraitId>,
    pub Shl: Option<TraitId>,
    pub Shr: Option<TraitId>,
    pub AddAssign: Option<TraitId>,
    pub SubAssign: Option<TraitId>,
    pub MulAssign: Option<TraitId>,
    pub DivAssign: Option<TraitId>,
    pub RemAssign: Option<TraitId>,
    pub BitXorAssign: Option<TraitId>,
    pub BitAndAssign: Option<TraitId>,
    pub BitOrAssign: Option<TraitId>,
    pub ShlAssign: Option<TraitId>,
    pub ShrAssign: Option<TraitId>,
    pub Index: Option<TraitId>,
    pub IndexMut: Option<TraitId>,
    pub UnsafeCell: Option<StructId>,
    pub UnsafePinned: Option<StructId>,
    pub VaList: Option<StructId>,
    pub Deref: Option<TraitId>,
    pub DerefMut: Option<TraitId>,
    pub DerefTarget: Option<TypeAliasId>,
    pub Receiver: Option<TraitId>,
    pub ReceiverTarget: Option<TypeAliasId>,
    pub Fn: Option<TraitId>,
    pub FnMut: Option<TraitId>,
    pub FnOnce: Option<TraitId>,
    pub AsyncFn: Option<TraitId>,
    pub AsyncFnMut: Option<TraitId>,
    pub AsyncFnOnce: Option<TraitId>,
    pub CallRefFuture: Option<TypeAliasId>,
    pub CallOnceFuture: Option<TypeAliasId>,
    pub AsyncFnOnceOutput: Option<TypeAliasId>,
    pub FnOnceOutput: Option<TypeAliasId>,
    pub Future: Option<TraitId>,
    pub CoroutineState: Option<EnumId>,
    pub Coroutine: Option<TraitId>,
    pub CoroutineReturn: Option<TypeAliasId>,
    pub CoroutineYield: Option<TypeAliasId>,
    pub Unpin: Option<TraitId>,
    pub Pin: Option<StructId>,
    pub PartialEq: Option<TraitId>,
    pub PartialOrd: Option<TraitId>,
    pub CVoid: Option<EnumId>,
    pub Panic: Option<FunctionId>,
    pub PanicNounwind: Option<FunctionId>,
    pub PanicFmt: Option<FunctionId>,
    pub PanicDisplay: Option<FunctionId>,
    pub ConstPanicFmt: Option<FunctionId>,
    pub PanicBoundsCheck: Option<FunctionId>,
    pub PanicMisalignedPointerDereference: Option<FunctionId>,
    pub PanicInfo: Option<StructId>,
    pub PanicLocation: Option<StructId>,
    pub PanicImpl: Option<FunctionId>,
    pub PanicCannotUnwind: Option<FunctionId>,
    pub PanicNullPointerDereference: Option<FunctionId>,
    pub BeginPanic: Option<FunctionId>,
    pub FormatAlignment: Option<EnumId>,
    pub FormatArgument: Option<StructId>,
    pub FormatArguments: Option<StructId>,
    pub FormatCount: Option<EnumId>,
    pub FormatPlaceholder: Option<StructId>,
    pub FormatUnsafeArg: Option<StructId>,
    pub ExchangeMalloc: Option<FunctionId>,
    pub BoxFree: Option<FunctionId>,
    pub DropInPlace: Option<FunctionId>,
    pub AllocLayout: Option<StructId>,
    pub Start: Option<FunctionId>,
    pub EhPersonality: Option<FunctionId>,
    pub EhCatchTypeinfo: Option<StaticId>,
    pub OwnedBox: Option<StructId>,
    pub PhantomData: Option<StructId>,
    pub ManuallyDrop: Option<StructId>,
    pub MaybeUninit: Option<UnionId>,
    pub AlignOffset: Option<FunctionId>,
    pub Termination: Option<TraitId>,
    pub Try: Option<TraitId>,
    pub Tuple: Option<TraitId>,
    pub SliceLen: Option<FunctionId>,
    pub TryTraitFromResidual: Option<FunctionId>,
    pub TryTraitFromOutput: Option<FunctionId>,
    pub TryTraitBranch: Option<FunctionId>,
    pub TryTraitFromYeet: Option<FunctionId>,
    pub ResidualIntoTryType: Option<FunctionId>,
    pub PointerLike: Option<TraitId>,
    pub ConstParamTy: Option<TraitId>,
    pub Poll: Option<EnumId>,
    pub PollReady: Option<EnumVariantId>,
    pub PollPending: Option<EnumVariantId>,
    pub ResumeTy: Option<StructId>,
    pub GetContext: Option<FunctionId>,
    pub Context: Option<StructId>,
    pub FuturePoll: Option<FunctionId>,
    pub FutureOutput: Option<TypeAliasId>,
    pub Option: Option<EnumId>,
    pub OptionSome: Option<EnumVariantId>,
    pub OptionNone: Option<EnumVariantId>,
    pub ResultOk: Option<EnumVariantId>,
    pub ResultErr: Option<EnumVariantId>,
    pub ControlFlowContinue: Option<EnumVariantId>,
    pub ControlFlowBreak: Option<EnumVariantId>,
    pub IntoFutureIntoFuture: Option<FunctionId>,
    pub IntoIterIntoIter: Option<FunctionId>,
    pub IteratorNext: Option<FunctionId>,
    pub Iterator: Option<TraitId>,
    pub PinNewUnchecked: Option<FunctionId>,
    pub RangeFrom: Option<StructId>,
    pub RangeFull: Option<StructId>,
    pub RangeInclusiveStruct: Option<StructId>,
    pub RangeInclusiveNew: Option<FunctionId>,
    pub Range: Option<StructId>,
    pub RangeToInclusive: Option<StructId>,
    pub RangeTo: Option<StructId>,
    pub RangeFromCopy: Option<StructId>,
    pub RangeInclusiveCopy: Option<StructId>,
    pub RangeCopy: Option<StructId>,
    pub RangeToInclusiveCopy: Option<StructId>,
    pub String: Option<StructId>,
    pub CStr: Option<StructId>,
    pub Ordering: Option<EnumId>,
    pub Default: Option<TraitId>,
    pub Debug: Option<TraitId>,
    pub Hash: Option<TraitId>,
    pub Ord: Option<TraitId>,
    pub Eq: Option<TraitId>,
    pub DefaultDerive: Option<MacroId>,
    pub DebugDerive: Option<MacroId>,
    pub HashDerive: Option<MacroId>,
    pub PartialOrdDerive: Option<MacroId>,
    pub OrdDerive: Option<MacroId>,
    pub PartialEqDerive: Option<MacroId>,
    pub EqDerive: Option<MacroId>,
    pub CoercePointeeDerive: Option<MacroId>,
    pub CopyDerive: Option<MacroId>,
    pub CloneDerive: Option<MacroId>,

}

Fields
Sized: Option<TraitId>MetaSized: Option<TraitId>PointeeSized: Option<TraitId>Unsize: Option<TraitId>StructuralPeq: Option<TraitId>

Trait injected by #[derive(PartialEq)], (i.e. Partial EQ).

StructuralTeq: Option<TraitId>

Trait injected by #[derive(Eq)], (i.e. Total EQ; no, I will not apologize).

Copy: Option<TraitId>Clone: Option<TraitId>TrivialClone: Option<TraitId>Sync: Option<TraitId>DiscriminantKind: Option<TraitId>Discriminant: Option<TypeAliasId>

The associated item of the DiscriminantKind trait.

PointeeTrait: Option<TraitId>Metadata: Option<TypeAliasId>DynMetadata: Option<StructId>Freeze: Option<TraitId>FnPtrTrait: Option<TraitId>FnPtrAddr: Option<FunctionId>Drop: Option<TraitId>Destruct: Option<TraitId>CoerceUnsized: Option<TraitId>DispatchFromDyn: Option<TraitId>TransmuteOpts: Option<StructId>TransmuteTrait: Option<TraitId>Add: Option<TraitId>Sub: Option<TraitId>Mul: Option<TraitId>Div: Option<TraitId>Rem: Option<TraitId>Neg: Option<TraitId>Not: Option<TraitId>BitXor: Option<TraitId>BitAnd: Option<TraitId>BitOr: Option<TraitId>Shl: Option<TraitId>Shr: Option<TraitId>AddAssign: Option<TraitId>SubAssign: Option<TraitId>MulAssign: Option<TraitId>DivAssign: Option<TraitId>RemAssign: Option<TraitId>BitXorAssign: Option<TraitId>BitAndAssign: Option<TraitId>BitOrAssign: Option<TraitId>ShlAssign: Option<TraitId>ShrAssign: Option<TraitId>Index: Option<TraitId>IndexMut: Option<TraitId>UnsafeCell: Option<StructId>UnsafePinned: Option<StructId>VaList: Option<StructId>Deref: Option<TraitId>DerefMut: Option<TraitId>DerefTarget: Option<TypeAliasId>Receiver: Option<TraitId>ReceiverTarget: Option<TypeAliasId>Fn: Option<TraitId>FnMut: Option<TraitId>FnOnce: Option<TraitId>AsyncFn: Option<TraitId>AsyncFnMut: Option<TraitId>AsyncFnOnce: Option<TraitId>CallRefFuture: Option<TypeAliasId>CallOnceFuture: Option<TypeAliasId>AsyncFnOnceOutput: Option<TypeAliasId>FnOnceOutput: Option<TypeAliasId>Future: Option<TraitId>CoroutineState: Option<EnumId>Coroutine: Option<TraitId>CoroutineReturn: Option<TypeAliasId>CoroutineYield: Option<TypeAliasId>Unpin: Option<TraitId>Pin: Option<StructId>PartialEq: Option<TraitId>PartialOrd: Option<TraitId>CVoid: Option<EnumId>Panic: Option<FunctionId>PanicNounwind: Option<FunctionId>PanicFmt: Option<FunctionId>PanicDisplay: Option<FunctionId>ConstPanicFmt: Option<FunctionId>PanicBoundsCheck: Option<FunctionId>PanicMisalignedPointerDereference: Option<FunctionId>PanicInfo: Option<StructId>PanicLocation: Option<StructId>PanicImpl: Option<FunctionId>PanicCannotUnwind: Option<FunctionId>PanicNullPointerDereference: Option<FunctionId>BeginPanic: Option<FunctionId>

libstd panic entry point. Necessary for const eval to be able to catch it

FormatAlignment: Option<EnumId>FormatArgument: Option<StructId>FormatArguments: Option<StructId>FormatCount: Option<EnumId>FormatPlaceholder: Option<StructId>FormatUnsafeArg: Option<StructId>ExchangeMalloc: Option<FunctionId>BoxFree: Option<FunctionId>DropInPlace: Option<FunctionId>AllocLayout: Option<StructId>Start: Option<FunctionId>EhPersonality: Option<FunctionId>EhCatchTypeinfo: Option<StaticId>OwnedBox: Option<StructId>PhantomData: Option<StructId>ManuallyDrop: Option<StructId>MaybeUninit: Option<UnionId>AlignOffset: Option<FunctionId>

Align offset for stride != 1; must not panic.

Termination: Option<TraitId>Try: Option<TraitId>Tuple: Option<TraitId>SliceLen: Option<FunctionId>TryTraitFromResidual: Option<FunctionId>TryTraitFromOutput: Option<FunctionId>TryTraitBranch: Option<FunctionId>TryTraitFromYeet: Option<FunctionId>ResidualIntoTryType: Option<FunctionId>PointerLike: Option<TraitId>ConstParamTy: Option<TraitId>Poll: Option<EnumId>PollReady: Option<EnumVariantId>PollPending: Option<EnumVariantId>ResumeTy: Option<StructId>GetContext: Option<FunctionId>Context: Option<StructId>FuturePoll: Option<FunctionId>FutureOutput: Option<TypeAliasId>Option: Option<EnumId>OptionSome: Option<EnumVariantId>OptionNone: Option<EnumVariantId>ResultOk: Option<EnumVariantId>ResultErr: Option<EnumVariantId>ControlFlowContinue: Option<EnumVariantId>ControlFlowBreak: Option<EnumVariantId>IntoFutureIntoFuture: Option<FunctionId>IntoIterIntoIter: Option<FunctionId>IteratorNext: Option<FunctionId>Iterator: Option<TraitId>PinNewUnchecked: Option<FunctionId>RangeFrom: Option<StructId>RangeFull: Option<StructId>RangeInclusiveStruct: Option<StructId>RangeInclusiveNew: Option<FunctionId>Range: Option<StructId>RangeToInclusive: Option<StructId>RangeTo: Option<StructId>RangeFromCopy: Option<StructId>RangeInclusiveCopy: Option<StructId>RangeCopy: Option<StructId>RangeToInclusiveCopy: Option<StructId>String: Option<StructId>CStr: Option<StructId>Ordering: Option<EnumId>Default: Option<TraitId>Debug: Option<TraitId>Hash: Option<TraitId>Ord: Option<TraitId>Eq: Option<TraitId>DefaultDerive: Option<MacroId>DebugDerive: Option<MacroId>HashDerive: Option<MacroId>PartialOrdDerive: Option<MacroId>OrdDerive: Option<MacroId>PartialEqDerive: Option<MacroId>EqDerive: Option<MacroId>CoercePointeeDerive: Option<MacroId>CopyDerive: Option<MacroId>CloneDerive: Option<MacroId>
Implementations

impl LangItems

fn collect_lang_item<T>(&mut self, db: &dyn DefDatabase, item: T)
where T: Into<AttrDefId> + Into<LangItemTarget> + Copy,

impl LangItems

fn is_empty(&self) -> bool

fn merge_prefer_self(&mut self, other: &Self)

Merges self with other, with preference to self items.

fn assign_lang_item(&mut self, name: Symbol, target: LangItemTarget)

fn fill_non_lang_core_items( &mut self, db: &dyn DefDatabase, core_def_map: &DefMap, )