hir_def::nameres

Module path_resolution

This modules implements a function to resolve a path foo::bar::baz to a def, which is used within the name resolution.

When name resolution is finished, the result of resolving a path is either Some(def) or None. However, when we are in process of resolving imports or macros, there’s a third possibility:

I can’t resolve this path right now, but I might be resolve this path later, when more macros are expanded.

ReachedFixedPoint signals about this.

Structs

ResolvePathResult 🔒

ResolvePathResultPrefixInfo

Enums

ReachedFixedPoint 🔒

ResolveMode 🔒

Functions

adjust_to_nearest_non_block_module 🔒

Given a block module, returns its nearest non-block module and the DefMap it belongs to.