hir_def

Module path

A desugared representation of paths like crate::foo or <Type as Trait>::bar.

Macros

path

Structs

AssociatedTypeBinding

An associated type binding like in Iterator<Item = T>.

GenericArgs

Generic arguments to a path segment (e.g. the i32 in Option<i32>). This also includes bindings of associated types, like in Iterator<Item = Foo>.

ModPath

NormalPath

PathSegment

PathSegments

Enums

GenericArg

A single generic argument.

GenericArgsParentheses

ImportAlias

Path

PathKind

Functions

hir_segment_to_ast_segment

This function finds the AST segment that corresponds to the HIR segment with index segment_idx on the path that is lowered from path.