hir_def
Module per_ns

In rust, it is possible to have a value, a type and a macro with the same name without conflicts.

PerNs (per namespace) captures this.

Structs

Item

PerNs

Enums

Namespace

Type Aliases

MacrosItem

TypesItem

ValuesItem