Trait hir_def::OpaqueInternableThing

pub trait OpaqueInternableThing: Any + Debug + Sync + Send + UnwindSafe + RefUnwindSafe {
    // Required methods
    fn as_any(&self) -> &dyn Any;
    fn box_any(&self) -> Box<dyn Any>;
    fn dyn_hash(&self, state: &mut dyn Hasher);
    fn dyn_eq(&self, other: &dyn OpaqueInternableThing) -> bool;
    fn dyn_clone(&self) -> Box<dyn OpaqueInternableThing>;
}

A thing that we want to store in interned ids, but we don’t know its type in hir-def. This is currently only used in InTypeConstId for storing the type (which has type Ty defined in the hir-ty crate) of the constant in its id, which is a temporary hack so we may want to remove this after removing that.

Required Methods

fn as_any(&self) -> &dyn Any

fn box_any(&self) -> Box<dyn Any>

fn dyn_hash(&self, state: &mut dyn Hasher)

fn dyn_eq(&self, other: &dyn OpaqueInternableThing) -> bool

fn dyn_clone(&self) -> Box<dyn OpaqueInternableThing>