hir_def

Type Alias ModuleId

pub type ModuleId = ModuleIdLt<'static>;

Aliased Type

pub struct ModuleId(pub(crate) Id, pub(crate) PhantomData<fn() -> &'static ()>);

Tuple Fields
0: Id1: PhantomData<fn() -> &'static ()>
Implementations

impl ModuleId

pub unsafe fn to_db<'db>(self, _db: &'db dyn DefDatabase) -> ModuleIdLt<'db>

Safety

The caller must ensure that the ModuleId comes from the given database.

pub fn def_map(self, db: &dyn DefDatabase) -> &DefMap

pub(crate) fn local_def_map( self, db: &dyn DefDatabase, ) -> (&DefMap, &LocalDefMap)

pub(crate) fn only_local_def_map(self, db: &dyn DefDatabase) -> &LocalDefMap

pub fn crate_def_map(self, db: &dyn DefDatabase) -> &DefMap

pub fn name(self, db: &dyn DefDatabase) -> Option<Name>

pub fn containing_module(self, db: &dyn DefDatabase) -> Option<ModuleId>

Returns the module containing self, either the parent mod, or the module (or block) containing the block, if self corresponds to a block expression.

pub fn is_block_module(self, db: &dyn DefDatabase) -> bool