Crate hir_def

hir_def crate contains everything between macro expansion and type inference.

It defines various items (structs, enums, traits) which comprises Rust code, as well as an algorithm for resolving paths to such entities.

Note that hir_def is a work in progress, so not all of the above is actually true.

Re-exports

pub use self::hir::type_ref;

pub use rustc_abi as layout;

pub use rustc_abi as layout;

pub use rustc_abi as layout;

Modules

attr

A higher level attributes based on TokenTree, with also some shortcuts.

body

Defines Body: a lowered representation of bodies of functions, statics and consts.

builtin_type

This module defines built-in types.

data

Contains basic data about various HIR declarations.

db

Defines database & queries for name resolution.

dyn_map

This module defines a DynMap – a container for heterogeneous maps.

expander

Macro expansion utilities.

find_path

An algorithm to find a path to refer to a certain item.

generics

Many kinds of items or constructs can have generic parameters: functions, structs, impls, traits, etc. This module provides a common HIR for these generic parameters. See also the Generics type and the generics_of query in rustc.

hir

This module describes hir-level representation of expressions.

import_map

A map of all publicly exported items in a crate.

item_scope

Describes items defined or visible (ie, imported) in a certain scope. This is shared between modules and blocks.

item_tree

A simplified AST that only contains items.

lang_item

Collects lang items: items marked with #[lang = "..."] attribute.

lower

Context for lowering paths.

nameres

This module implements import-resolution/macro expansion algorithm.

path

A desugared representation of paths like crate::foo or <Type as Trait>::bar.

per_ns

In rust, it is possible to have a value, a type and a macro with the same name without conflicts.

resolver

Name resolution façade.

src

Utilities for mapping between hir IDs and the surface syntax.

tt

visibility

Defines hir-level representation of visibility (e.g. pub and pub(crate)).

Structs

AssocItemLoc

BlockId

BlockLoc

ConstBlockId

Id of the anonymous const block expression and patterns. This is very similar to ClosureId and shouldn’t be a DefWithBodyId since its type inference is dependent on its parent.

ConstBlockLoc

ConstId

ConstParamId

A TypeOrConstParamId with an invariant that it actually belongs to a const

CrateRootModuleId

A ModuleId that is always a crate’s root module.

EnumId

EnumVariantId

EnumVariantLoc

ExternBlockId

ExternCrateId

FieldId

FunctionId

ImplId

ImportPathConfig

A wrapper around three booleans

InTypeConstId

A constant in a type as a substitution for const generics (like Foo<{ 2 + 2 }>) or as an array length (like [u8; 2 + 2]). These constants are body owner and are a variant of DefWithBodyId. These are not called AnonymousConstId to prevent confusion with ConstBlockId.

InTypeConstLoc

ItemLoc

LifetimeParamId

Macro2Id

Macro2Loc

MacroRulesId

MacroRulesLoc

MacroRulesLocFlags

ModuleId

ProcMacroId

ProcMacroLoc

StaticId

StructId

SyntheticSyntax

TraitAliasId

TraitId

TupleFieldId

TupleId

TypeAliasId

TypeOrConstParamId

TypeParamId

A TypeOrConstParamId with an invariant that it actually belongs to a type

UnionId

UnresolvedMacro

UseId

Enums

AdtId

A Data Type

AssocItemId

AttrDefId

CallableDefId

DefWithBodyId

The defs which have a body.

GeneralConstId

A constant, which might appears as a const item, an anonymous const block in expressions or patterns, or as a constant in types with const generics.

GenericDefId

GenericParamId

A generic param

ItemContainerId

MacroExpander

MacroId

A macro

ModuleDefId

The defs which can be visible in the module.

TypeOwnerId

Something that holds types, required for the current const arg lowering implementation as they need to be able to query where they are defined.

VariantId

Traits

AsMacroCall

A helper trait for converting to MacroCallId

HasModule

Intern

ItemTreeLoc

Lookup

OpaqueInternableThing

A thing that we want to store in interned ids, but we don’t know its type in hir-def. This is currently only used in InTypeConstId for storing the type (which has type Ty defined in the hir-ty crate) of the constant in its id, which is a temporary hack so we may want to remove this after removing that.

Type Aliases

EnumLoc

LocalFieldId

LocalLifetimeParamId

LocalModuleId

An ID of a module, local to a DefMap.

LocalTypeOrConstParamId

StaticLoc

TraitAliasLoc

TraitLoc

UnionLoc