Struct hir_def::body::Body

pub struct Body {
    pub exprs: Arena<Expr>,
    pub pats: Arena<Pat>,
    pub labels: Arena<Label>,
    pub params: Vec<PatId>,
    pub body_expr: ExprId,
    // some fields omitted
}

The body of an item (function, const etc.).

Fields
exprs: Arena<Expr>pats: Arena<Pat>labels: Arena<Label>params: Vec<PatId>

The patterns for the function’s parameters. While the parameter types are part of the function signature, the patterns are not (they don’t change the external type of the function).

If this Body is for the body of a constant, this will just be empty.

body_expr: ExprId

The ExprId of the actual body expression.

Implementations

impl Body

pub fn blocks<'a>(
&'a self,
db: &'a dyn DefDatabase
) -> impl Iterator<Item = (BlockId, Arc<DefMap>)> + '_

Returns an iterator over all block expressions in this body that define inner items.