hir_def::body
Struct BodySourceMap

pub struct BodySourceMap {
    pub types: TypesSourceMap,
    /* private fields */
}

An item body together with the mapping from syntax nodes to HIR expression IDs. This is needed to go from e.g. a position in a file to the HIR expression containing it; but for type inference etc., we want to operate on a structure that is agnostic to the actual positions of expressions in the file, so that we don’t recompute types whenever some whitespace is typed.

One complication here is that, due to macro expansion, a single Body might be spread across several files. So, for each ExprId and PatId, we record both the HirFileId and the position inside the file. However, we only store AST -> ExprId mapping for non-macro files, as it is not clear how to handle this properly for macros.

Fields
types: TypesSourceMap
Implementations

impl BodySourceMap

pub fn expr_or_pat_syntax( &self, id: ExprOrPatId, ) -> Result<ExprOrPatSource, SyntheticSyntax>

pub fn expr_syntax(&self, expr: ExprId) -> Result<ExprSource, SyntheticSyntax>

pub fn node_expr(&self, node: InFile<&Expr>) -> Option<ExprOrPatId>

pub fn node_macro_file(&self, node: InFile<&MacroCall>) -> Option<MacroFileId>

pub fn macro_calls( &self, ) -> impl Iterator<Item = (InFile<AstPtr<MacroCall>>, MacroFileId)> + '_

pub fn pat_syntax(&self, pat: PatId) -> Result<ExprOrPatSource, SyntheticSyntax>

pub fn self_param_syntax(&self) -> Option<InFile<AstPtr<SelfParam>>>

pub fn node_pat(&self, node: InFile<&Pat>) -> Option<PatId>

pub fn label_syntax(&self, label: LabelId) -> LabelSource

pub fn patterns_for_binding(&self, binding: BindingId) -> &[PatId]

pub fn node_label(&self, node: InFile<&Label>) -> Option<LabelId>

pub fn field_syntax(&self, expr: ExprId) -> FieldSource

pub fn pat_field_syntax(&self, pat: PatId) -> PatFieldSource

pub fn macro_expansion_expr( &self, node: InFile<&MacroExpr>, ) -> Option<ExprOrPatId>

pub fn expansions( &self, ) -> impl Iterator<Item = (&InFile<AstPtr<MacroCall>>, &MacroFileId)>

pub fn implicit_format_args( &self, node: InFile<&FormatArgsExpr>, ) -> Option<(HygieneId, &[(TextRange, Name)])>

pub fn format_args_implicit_capture( &self, capture_expr: ExprId, ) -> Option<InFile<(AstPtr<Expr>, TextRange)>>

pub fn asm_template_args( &self, node: InFile<&AsmExpr>, ) -> Option<(ExprId, &[Vec<(TextRange, usize)>])>

pub fn diagnostics(&self) -> &[BodyDiagnostic]

Get a reference to the body source map’s diagnostics.