Struct hir_def::body::BodySourceMap

pub struct BodySourceMap { /* fields omitted */ }

An item body together with the mapping from syntax nodes to HIR expression IDs. This is needed to go from e.g. a position in a file to the HIR expression containing it; but for type inference etc., we want to operate on a structure that is agnostic to the actual positions of expressions in the file, so that we don’t recompute types whenever some whitespace is typed.

One complication here is that, due to macro expansion, a single Body might be spread across several files. So, for each ExprId and PatId, we record both the HirFileId and the position inside the file. However, we only store AST -> ExprId mapping for non-macro files, as it is not clear how to handle this properly for macros.

Implementations

impl BodySourceMap

pub fn expr_syntax(&self, expr: ExprId) -> Result<ExprSource, SyntheticSyntax>

pub fn node_expr(&self, node: InFile<&Expr>) -> Option<ExprId>

pub fn node_macro_file(&self, node: InFile<&MacroCall>) -> Option<HirFileId>

pub fn pat_syntax(&self, pat: PatId) -> Result<PatSource, SyntheticSyntax>

pub fn node_pat(&self, node: InFile<&Pat>) -> Option<PatId>

pub fn node_self_param(&self, node: InFile<&SelfParam>) -> Option<PatId>

pub fn label_syntax(&self, label: LabelId) -> LabelSource

pub fn node_label(&self, node: InFile<&Label>) -> Option<LabelId>

pub fn field_syntax(&self, expr: ExprId) -> InFile<AstPtr<RecordExprField>>

pub fn node_field(&self, node: InFile<&RecordExprField>) -> Option<ExprId>