Module hir_def::builtin_attr

Builtin attributes resolved by nameres.

The actual definitions were copied from rustc’s compiler/rustc_feature/src/builtin_attrs.rs.

It was last synchronized with upstream commit c1a2db3372a4d6896744919284f3287650a38ab7.

The macros were adjusted to only expand to the attribute name, since that is all we need to do name resolution, and BUILTIN_ATTRIBUTES is almost entirely unchanged from the original, to ease updating.

Structs

AttributeTemplate

A template that the attribute input must match. Only top-level shape (#[attr] vs #[attr(...)] vs #[attr = ...]) is considered now.

BuiltinAttribute

Constants

INERT_ATTRIBUTES

“Inert” built-in attributes that have a special meaning to rustc or rustdoc.

TOOL_MODULES

Ignored attribute namespaces used by tools.

Functions

find_builtin_attr_idx