Module hir_def::hir

This module describes hir-level representation of expressions.

This representation is:

Identity-based. Each expression has an id, so we can distinguish between different 1 in 1 + 1.

Independent of syntax. Though syntactic provenance information can be attached separately via id-based side map.

Unresolved. Paths are stored as sequences of names, and not as defs the names refer to.

Desugared. There’s no if let.

See also a neighboring body module.

Modules

type_ref

HIR for references to types. Paths in these are not yet resolved. They can be directly created from an ast::TypeRef, without further queries.

Structs

Binding

FloatTypeWrapper

Label

MatchArm

RecordFieldPat

RecordLitField

Enums

ArithOp

Array

BinaryOp

BindingAnnotation

Explicit binding annotations given in the HIR for a binding. Note that this is not the final binding mode that we infer after type inference.

CaptureBy

ClosureKind

CmpOp

Expr

ExprOrPatId

Literal

LogicOp

Movability

Ordering

Pat

Close relative to rustc’s hir::PatKind

RangeOp

Statement

UnaryOp

Type Definitions

BindingId

ExprId

LabelId

PatId