Struct hir_def::item_scope::ItemScope

pub struct ItemScope { /* private fields */ }

Implementations

impl ItemScope

Legacy macros can only be accessed through special methods like get_legacy_macros. Other methods will only resolve values, types and module scoped macros only.

pub fn entries(&self) -> impl Iterator<Item = (&Name, PerNs)> + '_

pub fn imports(&self) -> impl Iterator<Item = ImportId> + '_

pub fn fully_resolve_import( &self, db: &dyn DefDatabase, import: ImportId ) -> PerNs

pub fn declarations(&self) -> impl Iterator<Item = ModuleDefId> + '_

pub fn extern_crate_decls( &self ) -> impl ExactSizeIterator<Item = ExternCrateId> + '_

pub fn use_decls(&self) -> impl ExactSizeIterator<Item = UseId> + '_

pub fn impls(&self) -> impl ExactSizeIterator<Item = ImplId> + '_

pub fn unnamed_consts(&self) -> impl Iterator<Item = ConstId> + '_

pub fn legacy_macros(&self) -> impl Iterator<Item = (&Name, &[MacroId])> + '_

Iterate over all legacy textual scoped macros visible at the end of the module

impl ItemScope

pub fn derive_macro_invoc( &self, ast_id: AstId<Adt>, attr_id: AttrId ) -> Option<MacroCallId>