Module hir_def::item_tree

A simplified AST that only contains items.

Structs

Const

Enum

ExternCrate

Field
A single field of an enum variant or struct

FileItemTreeId

Function

IdRange
A range of densely allocated ItemTree IDs.

Impl

Import
A desugared use import.

ItemTree
The item tree of a source file.

ItemTreeId

MacroCall

MacroDef
“Macros 2.0” macro definition.

MacroRules

Mod

RawVisibilityId

Static

Struct

Trait

TypeAlias

Union

Variant

Enums

AssocItem

AttrOwner

Fields

ModItem

ModKind

Param

StructDefKind

Traits

ItemTreeNode
Trait implemented by all item nodes in the item tree.