Trait hir_def::item_tree::ItemTreeNode

pub trait ItemTreeNode: Clone {
    type Source: AstNode + Into<Item>;

    fn ast_id(&self) -> FileAstId<Self::Source>;
    fn lookup(tree: &ItemTree, index: Idx<Self>) -> &Self;
    fn id_from_mod_item(mod_item: ModItem) -> Option<FileItemTreeId<Self>>;
    fn id_to_mod_item(id: FileItemTreeId<Self>) -> ModItem;
}

Trait implemented by all item nodes in the item tree.

Required Associated Types

type Source: AstNode + Into<Item>

Required Methods

fn ast_id(&self) -> FileAstId<Self::Source>

fn lookup(tree: &ItemTree, index: Idx<Self>) -> &Self

Looks up an instance of Self in an item tree.

fn id_from_mod_item(mod_item: ModItem) -> Option<FileItemTreeId<Self>>

Downcasts a ModItem to a FileItemTreeId specific to this type.

fn id_to_mod_item(id: FileItemTreeId<Self>) -> ModItem

Upcasts a FileItemTreeId to a generic ModItem.