Module hir_def::layout

Definitions needed for computing data layout of types.

Structs

AbiAndPrefAlign

A pair of alignments, ABI-mandated and preferred.

AddressSpace

An identifier that specifies the address space that some operation should operate on. Special address spaces have an effect on code generation, depending on the target and the address spaces it implements.

Align

Alignment of a type in bytes (always a power of two).

Niche

ReprFlags

ReprOptions

Represents the repr options provided by the user,

RustcEnumVariantIdx

Size

Size of a type in bytes.

TargetDataLayout

Parsed Data layout for a target, which contains everything needed to compute layouts.

WrappingRange

Inclusive wrap-around range of valid values, that is, if start > end, it represents start..=MAX, followed by 0..=end.

Enums

Abi

Describes how values of the type are passed by target ABIs, in terms of categories of C types there are ABI rules for.

Endian

Endianness of the target, which must match cfg(target-endian).

FieldsShape

Describes how the fields of a type are located in memory.

Integer

Integers, also used for enum discriminants.

IntegerType

LayoutError

Primitive

Fundamental unit of memory access and layout.

Scalar

Information about one scalar component of a Rust type.

StructKind

TargetDataLayoutErrors

Traits

IntegerExt

LayoutCalculator

Type Definitions

Layout

TagEncoding

Variants